[Refactor] transfer write / query Influx DB services into API rather than socket
