+++ server/routes/compiler.js
@@ -1,6 +1,8 @@
 const router = require('express').Router();
-const { runCompiler } = require('../controllers/compiler');
+const { runCompiler, writeRecord, queryRecord } = require('../controllers/compiler');
 
 router.get('/compiler', runCompiler);
+
+router.route('/record').get(queryRecord).post(writeRecord);
 
 module.exports = router;
